+++ tests/specs/mouse-controls/single-slider-vertical-test.js
@@ -77,9 +77,9 @@
       helper.fireMousedown(helper.slider.minH, 0, true);
 
       var expectedValue = 50,
-        offset = helper.slider.sliderElem.rzsp - helper.slider.valueToOffset(expectedValue) - helper.slider.handleHalfDim;
+        position = helper.slider.sliderElem.rzsp - helper.slider.valueToPosition(expectedValue) - helper.slider.handleHalfDim;
 
-      helper.fireMousemove(offset, true);
+      helper.fireMousemove(position, true);
       expect(helper.scope.slider.value).to.equal(expectedValue);
       helper.slider.positionTrackingHandle.called.should.be.true;
       helper.slider.callOnChange.called.should.be.true;
@@ -139,9 +139,9 @@
       sinon.spy(helper.slider, 'callOnChange');
 
       var expectedValue = 50,
-        offset = helper.slider.sliderElem.rzsp - helper.slider.valueToOffset(expectedValue) - helper.slider.handleHalfDim;
+        position = helper.slider.sliderElem.rzsp - helper.slider.valueToPosition(expectedValue) - helper.slider.handleHalfDim;
 
-      var event = helper.fireMousedown(helper.slider.fullBar, offset, true);
+      var event = helper.fireMousedown(helper.slider.fullBar, position, true);
 
       expect(helper.scope.slider.value).to.equal(expectedValue);
       expect(helper.slider.tracking).to.equal('lowValue');
@@ -156,9 +156,9 @@
       sinon.spy(helper.slider, 'callOnChange');
 
       var expectedValue = 12,
-        offset = helper.slider.sliderElem.rzsp - helper.slider.valueToOffset(expectedValue) - helper.slider.handleHalfDim;
+        position = helper.slider.sliderElem.rzsp - helper.slider.valueToPosition(expectedValue) - helper.slider.handleHalfDim;
 
-      helper.fireMousedown(helper.slider.selBar, offset, true);
+      helper.fireMousedown(helper.slider.selBar, position, true);
 
       expect(helper.scope.slider.value).to.equal(expectedValue);
       expect(helper.slider.tracking).to.equal('lowValue');
@@ -176,9 +176,9 @@
       sinon.spy(helper.slider, 'callOnChange');
 
       var expectedValue = 10,
-        offset = helper.slider.sliderElem.rzsp - helper.slider.valueToOffset(expectedValue) - helper.slider.handleHalfDim;
+        position = helper.slider.sliderElem.rzsp - helper.slider.valueToPosition(expectedValue) - helper.slider.handleHalfDim;
 
-      helper.fireMousedown(helper.slider.ticks, offset, true);
+      helper.fireMousedown(helper.slider.ticks, position, true);
 
       expect(helper.scope.slider.value).to.equal(expectedValue);
       expect(helper.slider.tracking).to.equal('lowValue');
@@ -196,9 +196,9 @@
       sinon.spy(helper.slider, 'callOnChange');
 
       var expectedValue = 10,
-        offset = helper.slider.sliderElem.rzsp - helper.slider.valueToOffset(expectedValue) - helper.slider.handleHalfDim;
+        position = helper.slider.sliderElem.rzsp - helper.slider.valueToPosition(expectedValue) - helper.slider.handleHalfDim;
 
-      helper.fireMousedown(helper.slider.ticks, offset, true);
+      helper.fireMousedown(helper.slider.ticks, position, true);
 
       expect(helper.scope.slider.value).to.equal(expectedValue);
       expect(helper.slider.tracking).to.equal('lowValue');
@@ -285,9 +285,9 @@
       helper.fireMousedown(helper.slider.minH, 0, true);
 
       var expectedValue = 50,
-        offset = helper.slider.sliderElem.rzsp - helper.slider.valueToOffset(expectedValue) - helper.slider.handleHalfDim;
+        position = helper.slider.sliderElem.rzsp - helper.slider.valueToPosition(expectedValue) - helper.slider.handleHalfDim;
 
-      helper.fireMousemove(offset, true);
+      helper.fireMousemove(position, true);
       expect(helper.scope.slider.value).to.equal(expectedValue);
       helper.slider.positionTrackingHandle.called.should.be.true;
       helper.slider.callOnChange.called.should.be.true;
@@ -347,9 +347,9 @@
       sinon.spy(helper.slider, 'callOnChange');
 
       var expectedValue = 50,
-        offset = helper.slider.sliderElem.rzsp - helper.slider.valueToOffset(expectedValue) - helper.slider.handleHalfDim;
+        position = helper.slider.sliderElem.rzsp - helper.slider.valueToPosition(expectedValue) - helper.slider.handleHalfDim;
 
-      var event = helper.fireMousedown(helper.slider.fullBar, offset, true);
+      var event = helper.fireMousedown(helper.slider.fullBar, position, true);
 
       expect(helper.scope.slider.value).to.equal(expectedValue);
       expect(helper.slider.tracking).to.equal('lowValue');
@@ -364,9 +364,9 @@
       sinon.spy(helper.slider, 'callOnChange');
 
       var expectedValue = 12,
-        offset = helper.slider.sliderElem.rzsp - helper.slider.valueToOffset(expectedValue) - helper.slider.handleHalfDim;
+        position = helper.slider.sliderElem.rzsp - helper.slider.valueToPosition(expectedValue) - helper.slider.handleHalfDim;
 
-      helper.fireMousedown(helper.slider.selBar, offset, true);
+      helper.fireMousedown(helper.slider.selBar, position, true);
 
       expect(helper.scope.slider.value).to.equal(expectedValue);
       expect(helper.slider.tracking).to.equal('lowValue');
@@ -384,9 +384,9 @@
       sinon.spy(helper.slider, 'callOnChange');
 
       var expectedValue = 10,
-        offset = helper.slider.sliderElem.rzsp - helper.slider.valueToOffset(expectedValue) - helper.slider.handleHalfDim;
+        position = helper.slider.sliderElem.rzsp - helper.slider.valueToPosition(expectedValue) - helper.slider.handleHalfDim;
 
-      helper.fireMousedown(helper.slider.ticks, offset, true);
+      helper.fireMousedown(helper.slider.ticks, position, true);
 
       expect(helper.scope.slider.value).to.equal(expectedValue);
       expect(helper.slider.tracking).to.equal('lowValue');
@@ -404,9 +404,9 @@
       sinon.spy(helper.slider, 'callOnChange');
 
       var expectedValue = 10,
-        offset = helper.slider.sliderElem.rzsp - helper.slider.valueToOffset(expectedValue) - helper.slider.handleHalfDim;
+        position = helper.slider.sliderElem.rzsp - helper.slider.valueToPosition(expectedValue) - helper.slider.handleHalfDim;
 
-      helper.fireMousedown(helper.slider.ticks, offset, true);
+      helper.fireMousedown(helper.slider.ticks, position, true);
 
       expect(helper.scope.slider.value).to.equal(expectedValue);
       expect(helper.slider.tracking).to.equal('lowValue');